refactor(tuners): relocate LHS-design into src/tuners/lhs_design/ subpackage
Mirror the pbt/ two-door layout so every strategy is a src/tuners/<strategy>/
package (tuner.py + cli.py + __main__.py + __init__.py), reachable both directly
(python -m src.tuners.lhs_design) and via the router (python -m src.tuners lhs).
Pure relocation, no behavior change.
